Refactor algorithms API to builder pattern
