+++ src/algo/shortest_paths.rs
@@ -1,29 +1,12 @@
 use std::borrow::Borrow;
-use std::cmp::{max, Reverse};
-use std::collections::{hash_map::Entry, BinaryHeap, HashSet};
-use std::hash::BuildHasherDefault;
 
-use rustc_hash::{FxHashMap, FxHashSet};
+use rustc_hash::FxHashMap;
 
-use crate::index::Virtual;
-use crate::infra::VisitSet;
-use crate::marker::{EdgeType, Outgoing};
-use crate::weight::Weighted;
-use crate::{traits::*, NumIndexType};
+use crate::traits::*;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum Algo {
-    Dijkstra,
-    BellmanFord,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Error {
-    NegativeWeight,
-    NegativeCycle,
-    EdgeNotAvailable,
-}
+mod bellman_ford;
+mod builder;
+mod dijkstra;
 
 #[derive(Debug)]
 pub struct ShortestPaths<W, G: GraphBase> {
@@ -33,93 +16,6 @@ pub struct ShortestPaths<W, G: GraphBase> {
     // subgraph which is significantly smaller than the original graph.
     dist: FxHashMap<G::VertexIndex, W>,
     pred: FxHashMap<G::VertexIndex, G::VertexIndex>,
-}
-
-impl<W, G> ShortestPaths<W, G>
-where
-    G: GraphBase,
-    W: Weight,
-{
-    pub fn run_algo<E, Ty: EdgeType, F>(
-        graph: &G,
-        start: G::VertexIndex,
-        goal: Option<G::VertexIndex>,
-        edge_weight: F,
-        algo: Option<Algo>,
-    ) -> Result<ShortestPaths<W, G>, Error>
-    where
-        G: VerticesBase + Edges<E, Ty> + VerticesBaseWeak + EdgesWeak<E, Ty> + Neighbors,
-        G::VertexIndex: NumIndexType,
-        F: Fn(&E) -> W,
-    {
-        let algo = algo.unwrap_or_else(|| {
-            if !W::is_unsigned() {
-                // There is a possibility that a negative weight is encountered,
-                // so we conservatively use Bellman-Ford.
-                Algo::BellmanFord
-            } else if goal.is_some() {
-                // If the goal is specified, Dijkstra's algorithm likely
-                // finishes without the need of traversing the entire graph.
-                Algo::Dijkstra
-            } else {
-                let v = graph.vertex_count();
-                let e = graph.edge_count();
-
-                // Compare the worst-case bounds. This will result in choosing
-                // Dijkstra in vast majority of cases.
-                if v * e < max(e, v * (v as f64).log2() as usize) {
-                    Algo::BellmanFord
-                } else {
-                    Algo::Dijkstra
-                }
-            }
-        });
-
-        match algo {
-            Algo::Dijkstra => dijkstra(graph, start, goal, edge_weight),
-            Algo::BellmanFord => bellman_ford(graph, start, edge_weight),
-        }
-    }
-
-    pub fn run<E, Ty: EdgeType, F>(
-        graph: &G,
-        start: G::VertexIndex,
-        goal: Option<G::VertexIndex>,
-        edge_weight: F,
-    ) -> Result<ShortestPaths<W, G>, Error>
-    where
-        G: VerticesBase + Edges<E, Ty> + VerticesBaseWeak + EdgesWeak<E, Ty> + Neighbors,
-        G::VertexIndex: NumIndexType,
-        F: Fn(&E) -> W,
-    {
-        Self::run_algo(graph, start, goal, edge_weight, None)
-    }
-
-    pub fn run_dijkstra<E, Ty: EdgeType, F>(
-        graph: &G,
-        start: G::VertexIndex,
-        goal: Option<G::VertexIndex>,
-        edge_weight: F,
-    ) -> Result<ShortestPaths<W, G>, Error>
-    where
-        G: VerticesBaseWeak + EdgesWeak<E, Ty> + Neighbors,
-        F: Fn(&E) -> W,
-    {
-        dijkstra(graph, start, goal, edge_weight)
-    }
-
-    pub fn run_bellman_ford<E, Ty: EdgeType, F>(
-        graph: &G,
-        start: G::VertexIndex,
-        edge_weight: F,
-    ) -> Result<ShortestPaths<W, G>, Error>
-    where
-        G: VerticesBase + Edges<E, Ty>,
-        G::VertexIndex: NumIndexType,
-        F: Fn(&E) -> W,
-    {
-        bellman_ford(graph, start, edge_weight)
-    }
 }
 
 impl<W, G> ShortestPaths<W, G>
@@ -146,6 +42,36 @@ where
     }
 }
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[non_exhaustive]
+pub enum Algo {
+    Dijkstra,
+    BellmanFord,
+}
+
+mod algo {
+    use super::Algo;
+
+    #[derive(Debug)]
+    pub struct Any;
+
+    #[derive(Debug)]
+    pub struct Dijkstra;
+
+    #[derive(Debug)]
+    pub struct BellmanFord;
+
+    #[derive(Debug)]
+    pub struct Specific(pub Option<Algo>);
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Error {
+    NegativeWeight,
+    NegativeCycle,
+    EdgeNotAvailable,
+}
+
 pub struct PathReconstruction<'a, G: GraphBase> {
     curr: G::VertexIndex,
     pred: &'a FxHashMap<G::VertexIndex, G::VertexIndex>,
@@ -158,194 +84,6 @@ impl<'a, G: GraphBase> Iterator for PathReconstruction<'a, G> {
         self.curr = self.pred.get(&self.curr).cloned()?;
         Some(self.curr.clone())
     }
-}
-
-pub fn identity<E: Clone>(edge: &E) -> E {
-    edge.clone()
-}
-
-pub fn unit<E>(_edge: &E) -> usize {
-    1
-}
-
-fn dijkstra<E, Ty: EdgeType, G, W, F>(
-    graph: &G,
-    start: G::VertexIndex,
-    goal: Option<G::VertexIndex>,
-    edge_weight: F,
-) -> Result<ShortestPaths<W, G>, Error>
-where
-    G: VerticesBaseWeak + EdgesWeak<E, Ty> + Neighbors,
-    W: Weight,
-    F: Fn(&E) -> W,
-{
-    // Not using FixedBitSet with CompactIndexMap because the algorithm supports
-    // early termination when reaching given goal. It is likely that reaching
-    // goal means visiting a subgraph which is significantly smaller than the
-    // original graph.
-    let mut visited: FxHashSet<_> = HashSet::with_capacity_and_hasher(
-        graph.vertex_count_hint().unwrap_or(32),
-        BuildHasherDefault::default(),
-    );
-
-    let mut dist = FxHashMap::default();
-    let mut pred = FxHashMap::default();
-    let mut queue = BinaryHeap::new();
-
-    dist.insert(start.clone(), W::zero());
-    queue.push(Reverse(Weighted(start.clone(), W::zero())));
-
-    while let Some(Reverse(Weighted(vertex, vertex_dist))) = queue.pop() {
-        // This can happen due to duplication of vertices when doing relaxation
-        // in our implementation.
-        if visited.is_visited(&vertex) {
-            continue;
-        }
-
-        if goal.as_ref() == Some(&vertex) {
-            break;
-        }
-
-        for neighbor in graph.neighbors_directed(&vertex, Outgoing) {
-            let edge = graph
-                .edge_weak(&neighbor.edge())
-                .ok_or(Error::EdgeNotAvailable)?;
-            let next = neighbor.index().into_owned();
-
-            if visited.is_visited(&next) {
-                continue;
-            }
-
-            let edge_dist = edge_weight(&edge);
-
-            // The check for unsignedness should eliminate the negativity weight
-            // check, because the implementation of `is_unsigned` method is
-            // always a constant boolean in practice.
-            if !W::is_unsigned() && edge_dist < W::zero() {
-                return Err(Error::NegativeWeight);
-            }
-
-            let next_dist = vertex_dist.clone() + edge_dist;
-
-            match dist.entry(next.clone()) {
-                Entry::Occupied(curr_dist) => {
-                    // Relaxation operation. If the distance is better than what
-                    // we had so far, update it.
-                    if next_dist < *curr_dist.get() {
-                        *curr_dist.into_mut() = next_dist.clone();
-                        // A textbook version of the algorithm would update the
-                        // priority of `next`. Adding it as a new item causes
-                        // duplicities which is unfortunate for dense graphs,
-                        // but should be fine in practice.
-                        queue.push(Reverse(Weighted(next.clone(), next_dist)));
-                        pred.insert(next, vertex.clone());
-                    }
-                }
-                Entry::Vacant(slot) => {
-                    slot.insert(next_dist.clone());
-                    queue.push(Reverse(Weighted(next.clone(), next_dist)));
-                    pred.insert(next, vertex.clone());
-                }
-            }
-
-            // The vertex is finished.
-            visited.visit(vertex.clone());
-        }
-    }
-
-    Ok(ShortestPaths { start, dist, pred })
-}
-
-fn bellman_ford<E, Ty: EdgeType, G, W, F>(
-    graph: &G,
-    start: G::VertexIndex,
-    edge_weight: F,
-) -> Result<ShortestPaths<W, G>, Error>
-where
-    G: VerticesBase + Edges<E, Ty>,
-    G::VertexIndex: NumIndexType,
-    W: Weight,
-    F: Fn(&E) -> W,
-{
-    let vertex_map = graph.vertex_index_map();
-
-    let mut dist = vec![W::inf(); vertex_map.len()];
-    let mut pred = vec![Virtual::null(); vertex_map.len()];
-
-    dist[vertex_map.virt(start).unwrap().to_usize()] = W::zero();
-
-    let mut terminated_early = false;
-
-    // Try to relax edges |V| - 1 times.
-    for _ in 1..graph.vertex_count() {
-        let mut relaxed = false;
-
-        for edge in graph.edges() {
-            let u = vertex_map.virt(*edge.src()).unwrap();
-            let v = vertex_map.virt(*edge.dst()).unwrap();
-
-            let edge_dist = edge_weight(edge.data());
-            let next_dist = dist[u.to_usize()].clone() + edge_dist;
-
-            // Relax if better.
-            if next_dist < dist[v.to_usize()] {
-                dist[v.to_usize()] = next_dist;
-                pred[v.to_usize()] = u;
-                relaxed = true;
-            }
-        }
-
-        // If no distance was improved, then subsequent iterations would not
-        // improve as well. So we can terminate early.
-        if !relaxed {
-            terminated_early = true;
-            break;
-        }
-    }
-
-    // Check for negative cycles. If the main loop was terminated early, then
-    // the absence of cycle if guaranteed.
-    if !terminated_early {
-        for edge in graph.edges() {
-            let u = vertex_map.virt(*edge.src()).unwrap();
-            let v = vertex_map.virt(*edge.dst()).unwrap();
-
-            let edge_dist = edge_weight(edge.data());
-
-            if dist[u.to_usize()].clone() + edge_dist < dist[v.to_usize()] {
-                return Err(Error::NegativeCycle);
-            }
-        }
-    }
-
-    let dist = dist
-        .into_iter()
-        .enumerate()
-        .filter_map(|(i, d)| {
-            if d != W::inf() {
-                vertex_map.real(Virtual::from(i)).map(|u| (u, d))
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    let pred = pred
-        .into_iter()
-        .enumerate()
-        .filter_map(|(i, p)| {
-            if !p.is_null() {
-                Some((
-                    vertex_map.real(Virtual::from(i)).unwrap(),
-                    vertex_map.real(p).unwrap(),
-                ))
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    Ok(ShortestPaths { start, dist, pred })
 }
 
 #[cfg(test)]
@@ -382,7 +120,10 @@ mod tests {
     #[test]
     fn dijkstra_basic() {
         let graph = create_basic_graph();
-        let shortest_paths = ShortestPaths::run_dijkstra(&graph, 0.into(), None, identity).unwrap();
+        let shortest_paths = ShortestPaths::on(&graph)
+            .with(Algo::Dijkstra)
+            .run(0.into())
+            .unwrap();
 
         assert_eq!(shortest_paths.dist(&4.into()), Some(&8));
         assert_eq!(
@@ -396,8 +137,11 @@ mod tests {
     #[test]
     fn dijkstra_early_termination() {
         let graph = create_basic_graph();
-        let shortest_paths =
-            ShortestPaths::run_dijkstra(&graph, 0.into(), Some(4.into()), identity).unwrap();
+        let shortest_paths = ShortestPaths::on(&graph)
+            .goal(4.into())
+            .with(Algo::Dijkstra)
+            .run(0.into())
+            .unwrap();
 
         assert!(shortest_paths.dist(&5.into()).is_none());
     }
@@ -407,8 +151,10 @@ mod tests {
         let mut graph = create_basic_graph();
         graph.replace_edge(&2.into(), -1);
 
-        let shortest_paths =
-            ShortestPaths::run_dijkstra(&graph, 0.into(), Some(4.into()), identity);
+        let shortest_paths = ShortestPaths::on(&graph)
+            .goal(4.into())
+            .with(Algo::Dijkstra)
+            .run(0.into());
 
         assert_matches!(shortest_paths, Err(Error::NegativeWeight));
     }
@@ -416,7 +162,10 @@ mod tests {
     #[test]
     fn bellman_ford_basic() {
         let graph = create_basic_graph();
-        let shortest_paths = ShortestPaths::run_bellman_ford(&graph, 0.into(), identity).unwrap();
+        let shortest_paths = ShortestPaths::on(&graph)
+            .with(Algo::BellmanFord)
+            .run(0.into())
+            .unwrap();
 
         assert_eq!(shortest_paths.dist(&4.into()), Some(&8));
         assert_eq!(
@@ -432,7 +181,10 @@ mod tests {
         let mut graph = create_basic_graph();
         graph.replace_edge(&2.into(), -1);
 
-        let shortest_paths = ShortestPaths::run_bellman_ford(&graph, 0.into(), identity).unwrap();
+        let shortest_paths = ShortestPaths::on(&graph)
+            .with(Algo::BellmanFord)
+            .run(0.into())
+            .unwrap();
 
         assert_eq!(shortest_paths.dist(&4.into()), Some(&8));
         assert_eq!(
@@ -459,7 +211,9 @@ mod tests {
         graph.add_edge(&v2, &v1, -2);
         graph.add_edge(&v2, &v4, 3);
 
-        let shortest_paths = ShortestPaths::run_bellman_ford(&graph, 0.into(), identity);
+        let shortest_paths = ShortestPaths::on(&graph)
+            .with(Algo::BellmanFord)
+            .run(0.into());
 
         assert_matches!(shortest_paths, Err(Error::NegativeCycle));
     }
