+++ src/algo/toposort.rs
@@ -1,70 +1,17 @@
-use std::hash::BuildHasherDefault;
+use crate::{index::NumIndexType, marker::Directed, traits::*, visit};
 
-use rustc_hash::FxHashSet;
+mod builder;
+mod dfs;
+mod kahn;
 
-use crate::{
-    index::{NumIndexType, UseVertexIndex},
-    infra::{CompactIndexMap, VisitSet},
-    marker::{Directed, Direction},
-    operator::Transpose,
-    traits::*,
-    visit::{self, raw::*, VisitAll, Visitor},
-};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum Algo {
-    Dfs,
-    Kahn,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Error {
-    Cycle,
-}
+pub use dfs::DfsVisit;
+use kahn::KahnIter;
 
 pub struct TopoSort<'a, G>
 where
     G: VerticesBase + EdgesBase<Directed>,
 {
     inner: TopoSortInner<'a, G>,
-}
-
-impl<'a, G> TopoSort<'a, G>
-where
-    G: Neighbors + VerticesBase + EdgesBase<Directed>,
-{
-    pub fn run_algo(graph: &'a G, algo: Option<Algo>) -> Self
-    where
-        G::VertexIndex: NumIndexType,
-    {
-        match algo {
-            Some(Algo::Dfs) | None => Self {
-                inner: TopoSortInner::Dfs(Self::run_dfs(graph).into_iter(graph)),
-            },
-            Some(Algo::Kahn) => Self {
-                inner: TopoSortInner::Kahn(Self::run_kahn(graph)),
-            },
-        }
-    }
-
-    pub fn run(graph: &'a G) -> Self
-    where
-        G::VertexIndex: NumIndexType,
-    {
-        Self::run_algo(graph, None)
-    }
-
-    pub fn run_dfs(graph: &'a G) -> DfsVisit<'a, G> {
-        DfsVisit::new(graph)
-    }
-
-    pub fn run_kahn(graph: &'a G) -> KahnIter<'a, G>
-    where
-        G::VertexIndex: NumIndexType,
-    {
-        KahnIter::new(graph)
-    }
 }
 
 impl<'a, G> Iterator for TopoSort<'a, G>
@@ -77,6 +24,34 @@ where
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[non_exhaustive]
+pub enum Algo {
+    Dfs,
+    Kahn,
+}
+
+mod algo {
+    use super::Algo;
+
+    #[derive(Debug)]
+    pub struct Any;
+
+    #[derive(Debug)]
+    pub struct Dfs;
+
+    #[derive(Debug)]
+    pub struct Kahn;
+
+    #[derive(Debug)]
+    pub struct Specific(pub Option<Algo>);
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Error {
+    Cycle,
 }
 
 enum TopoSortInner<'a, G>
@@ -102,156 +77,6 @@ where
     }
 }
 
-pub struct DfsVisit<'a, G>
-where
-    G: VerticesBase + 'a,
-{
-    raw: RawVisit<G, UseVertexIndex, RawDfsExtra>,
-    multi: RawVisitMulti<G, UseVertexIndex, RawDfsExtra, VisitAll<'a, G>>,
-    closed: FxHashSet<G::VertexIndex>,
-}
-
-impl<'a, G> DfsVisit<'a, G>
-where
-    G: VerticesBase + 'a,
-{
-    fn new(graph: &'a G) -> Self {
-        Self {
-            raw: RawVisit::new(Some(graph.vertex_count())),
-            multi: RawVisitMulti::new(VisitAll::new(graph)),
-            closed: FxHashSet::with_capacity_and_hasher(
-                graph.vertex_count(),
-                BuildHasherDefault::default(),
-            ),
-        }
-    }
-}
-
-impl<'a, G> Visitor<G> for DfsVisit<'a, G>
-where
-    G: Neighbors + VerticesBase + EdgesBase<Directed>,
-{
-    type Item = Result<G::VertexIndex, Error>;
-
-    fn next(&mut self, graph: &G) -> Option<Self::Item> {
-        // The implementation differs from classic DFS algorithm for topological
-        // sorting in order to achieve lazy behavior. The algorithm traverses
-        // the graph in reverse DFS order and reports each vertex that is being
-        // closed (i.e., when we reached vertex that has no (unreported)
-        // predecessors). If a back edge is encountered, the cycle error is
-        // reported instead.
-
-        // Reverse the directions of the edges so the traversal actually works
-        // in reverse direction.
-        let graph = &Transpose::new(graph);
-
-        loop {
-            let mut cycle = false;
-
-            let raw_extra_event = self.multi.next_multi(
-                &mut self.raw,
-                |raw| {
-                    raw.next(graph, |_, raw_event| match raw_event {
-                        RawEvent::Skip { vertex, .. } if !self.closed.is_visited(&vertex) => {
-                            // Vertex not added to the stack, but also
-                            // has not been closed. That means that we
-                            // encountered a back edge.
-                            cycle = true;
-
-                            // Prune to avoid unnecessary work.
-                            false
-                        }
-                        _ => true,
-                    })
-                },
-                |vertex| graph.contains_vertex(vertex),
-            )?;
-
-            if cycle {
-                return Some(Err(Error::Cycle));
-            }
-
-            if let RawDfsExtraEvent::Close(vertex) = raw_extra_event {
-                self.closed.visit(vertex.clone());
-                return Some(Ok(vertex));
-            }
-        }
-    }
-}
-
-pub struct KahnIter<'a, G>
-where
-    G: GraphBase,
-{
-    graph: &'a G,
-    map: CompactIndexMap<G::VertexIndex>,
-    in_deg: Vec<usize>,
-    // Does not need to be FIFO as the order of reported vertices with in degree
-    // 0 does not matter.
-    queue: Vec<G::VertexIndex>,
-    visited: usize,
-}
-
-impl<'a, G> KahnIter<'a, G>
-where
-    G: Neighbors + VerticesBase + 'a,
-    G::VertexIndex: NumIndexType,
-{
-    fn new(graph: &'a G) -> Self {
-        let map = graph.vertex_index_map();
-        let mut in_deg = Vec::with_capacity(map.len());
-        let mut queue = Vec::new();
-
-        for v in graph.vertex_indices() {
-            let deg = graph.degree_directed(&v, Direction::Incoming);
-            in_deg.push(deg);
-
-            if deg == 0 {
-                queue.push(v);
-            }
-        }
-
-        Self {
-            graph,
-            map,
-            in_deg,
-            queue,
-            visited: 0,
-        }
-    }
-}
-
-impl<'a, G> Iterator for KahnIter<'a, G>
-where
-    G: Neighbors,
-    G::VertexIndex: NumIndexType,
-{
-    type Item = Result<G::VertexIndex, Error>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some(vertex) = self.queue.pop() {
-            self.visited += 1;
-
-            for n in self.graph.neighbors_directed(&vertex, Direction::Outgoing) {
-                let i = self.map.virt(*n.index()).unwrap().to_usize();
-                let deg = &mut self.in_deg[i];
-                *deg -= 1;
-
-                if *deg == 0 {
-                    self.queue.push(*n.index());
-                }
-            }
-
-            Some(Ok(vertex))
-        } else if self.visited != self.map.len() {
-            // `self.map.len()` corresponds to vertex count.
-            Some(Err(Error::Cycle))
-        } else {
-            None
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -259,7 +84,7 @@ mod tests {
     use crate::{
         index::DefaultIndexing,
         storage::AdjList,
-        visit::{DfsEvent, DfsEvents},
+        visit::{DfsEvent, DfsEvents, Visitor},
     };
 
     fn assert_valid<'a, G>(toposort: TopoSort<'a, G>, graph: &'a G)
@@ -388,7 +213,7 @@ mod tests {
     #[test]
     fn dfs_basic() {
         let graph = create_basic_graph();
-        let toposort = TopoSort::run_algo(&graph, Some(Algo::Dfs));
+        let toposort = TopoSort::on(&graph).with(Algo::Dfs).run();
 
         assert_valid(toposort, &graph);
     }
@@ -396,7 +221,7 @@ mod tests {
     #[test]
     fn dfs_cycle() {
         let graph = create_cyclic_graph();
-        let toposort = TopoSort::run_algo(&graph, Some(Algo::Dfs));
+        let toposort = TopoSort::on(&graph).with(Algo::Dfs).run();
 
         assert_valid(toposort, &graph);
     }
@@ -404,7 +229,7 @@ mod tests {
     #[test]
     fn dfs_disconnected() {
         let graph = create_disconnected_graph();
-        let toposort = TopoSort::run_algo(&graph, Some(Algo::Dfs));
+        let toposort = TopoSort::on(&graph).with(Algo::Dfs).run();
 
         assert_valid(toposort, &graph);
     }
@@ -412,7 +237,7 @@ mod tests {
     #[test]
     fn kahn_basic() {
         let graph = create_basic_graph();
-        let toposort = TopoSort::run_algo(&graph, Some(Algo::Kahn));
+        let toposort = TopoSort::on(&graph).with(Algo::Kahn).run();
 
         assert_valid(toposort, &graph);
     }
@@ -420,7 +245,7 @@ mod tests {
     #[test]
     fn kahn_cycle() {
         let graph = create_cyclic_graph();
-        let toposort = TopoSort::run_algo(&graph, Some(Algo::Kahn));
+        let toposort = TopoSort::on(&graph).with(Algo::Kahn).run();
 
         assert_valid(toposort, &graph);
     }
@@ -428,7 +253,7 @@ mod tests {
     #[test]
     fn kahn_disconnected() {
         let graph = create_disconnected_graph();
-        let toposort = TopoSort::run_algo(&graph, Some(Algo::Kahn));
+        let toposort = TopoSort::on(&graph).with(Algo::Kahn).run();
 
         assert_valid(toposort, &graph);
     }
